add application image with vault and consul

+++ build-image.sh
@@ -73,7 +73,9 @@ echo "Found latest Image in Family ${image_family}: $(echo ${latest_image_raw} |
 export PKR_VAR_clone_vm_id=$(echo ${latest_image_raw} | jq -r '.id | sub("^qemu/"; "")')
 
 # Build the new image
-ansible-galaxy install -r ${image_path}/ansible/requirements.yml --force --keep-scm-meta
+if [ -f "${image_path}/ansible/requirements.yml" ]; then
+  ansible-galaxy install -r ${image_path}/ansible/requirements.yml --force --keep-scm-meta
+fi
 (cd ${image_path} && packer init main.pkr.hcl && packer build -force -timestamp-ui main.pkr.hcl)
 
 # TODO: during PRs we need the option to not continue and delete the created vm
